+++ packages/frontend/src/components/index.ts
@@ -1,3 +1,4 @@
+import AuthComponent from './Auth';
 import Dropdown from './Dropdown';
 import Modal from './Modal';
-export { Dropdown, Modal };
+export { AuthComponent, Dropdown, Modal };
